The mapping of error line numbers returned by the SQL Server exceptions to the line numbers in the source file was sometimes wrong. When the current batch was produced by the preprocessor csql file and the line number at the beginning of the batch to calculated the output. It did not respect the possibility that the context in one batch could change. This can happen either because of multi line macro expansions, included source in the current batch or when using the pre processor option to remove empty lines.
This bug is solved now. Test cases have been added to fixture the new behaviour.

Furthermore a first step to refactor the main class has been taken. The class Processor mixes up to many different aspects and needs to be splitted for better separation of concerns. It will be the matter of another task to finish it.